Move settings to user level
